requirements.txt
flask
onnxruntime
pillow
numpy
gunicorn
cv2